refactor: extract shared logic
